+++ packages/ipns/src/routing/index.ts
@@ -1,7 +1,7 @@
 import type { DHTProgressEvents } from './dht.js'
 import type { DatastoreProgressEvents } from './local-store.js'
 import type { PubSubProgressEvents } from './pubsub.js'
-import type { AbortOptions } from '@libp2p/interfaces'
+import type { AbortOptions } from '@libp2p/interface'
 import type { ProgressOptions } from 'progress-events'
 
 export interface PutOptions extends AbortOptions, ProgressOptions {
@@ -13,8 +13,8 @@ export interface GetOptions extends AbortOptions, ProgressOptions {
 }
 
 export interface IPNSRouting {
-  put: (routingKey: Uint8Array, marshaledRecord: Uint8Array, options?: PutOptions) => Promise<void>
-  get: (routingKey: Uint8Array, options?: GetOptions) => Promise<Uint8Array>
+  put(routingKey: Uint8Array, marshaledRecord: Uint8Array, options?: PutOptions): Promise<void>
+  get(routingKey: Uint8Array, options?: GetOptions): Promise<Uint8Array>
 }
 
 export type IPNSRoutingEvents =