+++ packages/ipns/src/routing/local-store.ts
@@ -1,9 +1,9 @@
-import { Libp2pRecord } from '@libp2p/record'
+import { Record } from '@libp2p/kad-dht'
 import { type Datastore, Key } from 'interface-datastore'
 import { CustomProgressEvent, type ProgressEvent } from 'progress-events'
 import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
 import type { GetOptions, IPNSRouting, PutOptions } from '../routing'
-import type { AbortOptions } from '@libp2p/interfaces'
+import type { AbortOptions } from '@libp2p/interface'
 
 function dhtRoutingKey (key: Uint8Array): Key {
   return new Key('/dht/record/' + uint8ArrayToString(key, 'base32'), false)
@@ -15,7 +15,7 @@ export type DatastoreProgressEvents =
   ProgressEvent<'ipns:routing:datastore:error', Error>
 
 export interface LocalStore extends IPNSRouting {
-  has: (routingKey: Uint8Array, options?: AbortOptions) => Promise<boolean>
+  has(routingKey: Uint8Array, options?: AbortOptions): Promise<boolean>
 }
 
 /**
@@ -30,7 +30,7 @@ export function localStore (datastore: Datastore): LocalStore {
         const key = dhtRoutingKey(routingKey)
 
         // Marshal to libp2p record as the DHT does
-        const record = new Libp2pRecord(routingKey, marshalledRecord, new Date())
+        const record = new Record(routingKey, marshalledRecord, new Date())
 
         options.onProgress?.(new CustomProgressEvent('ipns:routing:datastore:put'))
         await datastore.put(key, record.serialize(), options)
@@ -47,7 +47,7 @@ export function localStore (datastore: Datastore): LocalStore {
         const buf = await datastore.get(key, options)
 
         // Unmarshal libp2p record as the DHT does
-        const record = Libp2pRecord.deserialize(buf)
+        const record = Record.deserialize(buf)
 
         return record.value
       } catch (err: any) {